+++ CHANGES.md
@@ -2,6 +2,7 @@
 
 | CamFlow version | Kernel version | Date       |
 | --------------- |----------------| ---------- |
+| 0.3.11          | 4.14.18        | 09/02/2018 |
 | 0.3.10          | 4.14.15        | 27/01/2018 |
 | 0.3.9           | 4.14.9         | 27/12/2017 |
 | 0.3.8           | 4.14.5         | 12/12/2017 |
@@ -29,6 +30,13 @@
 | 0.1.2           | 4.4.6          | 26/05/2016 |
 | 0.1.1           | 4.4.6          | 03/04/2016 |
 | 0.1.0           | 4.2.8          | 28/03/2016 |
+
+### v0.3.11
+```
+- RL_SND_UNIX and RL_RCV_UNIX replace by simply RL_SND and RL_RCV.
+- Fix issue with user buffer in fs interface (only affected certain configuration).
+- Update to kernel version 4.14.18.
+```
 
 ### v0.3.10
 ```